Allow for any arguments to be passed to the lambda function
Before, this was limited to lambda functions that are expecting `message` in the
events map.  However this assumption is not universal, and as such, should allow
the user to pass any arguments they want.

For example..

`hubot lambda function-name arg1:Yo arg2:Ho`

will send the following events dict to the lambda function

`{'arg1': 'Yo', 'arg2': 'Ho'}`